samples/IoT-Edge/Python3/TpmSample/requirements.txt
requests==2.28.1
flask==2.2.2
azure-iot-device==2.11.0